Organize all pages into their own components, and add proper typing
