+++ src/components/ServicePage/SolutionsSection/SolutionComponent.tsx
